Make issue_sidebar_spec.rb less flaky
The previous behavior attempted to click on some arbitrary point of the
sidebar to hide the sidebar, but in many occasions this point was
obscured by the opened assignee dropdown already.

To make this spec less flaky, click on the `Participants` section to
make the sidebar go away.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/329662