Add Experiment Surrounding Environment Guidance
We want to be able to limit the amount of users who can see this alert
as a slow rollout to see if it is useful.

This experiment is only enabled for projects that don't have existing
environments, and we currently only track if the user clicks on the
Learn More link embedded in the alert.

As with all experiments, this is only applicable to GitLab.com