Bust the cache for /whats-new
Using a query param to bust the cache on requests to /whats-new. While I
haven't personally experienced a cacheing issue, it has been reported
that the count and notification bubble are correct, though when loading
the "What's New" module, the content is stale.

Initially I combed over the cache headers, thinking maybe something was
missed there, but the headers seem correct. And given I can't reproduce
the experience, this issue is difficult to solve.

I chose this simple approach to see if it solves the issue, though I'm
open to other solutions.

part of:
https://gitlab.com/gitlab-org/gitlab/-/issues/325899