Remove user from rotation if removed from project/group
This adds a service to remove a user from an On-call Rotation.

We ensure that the rotation is up to date by running the job
top persist current shifts. Then, we update the `is_removed`
boolean on the Participant model associated to the
user/rotation.

Deletion from rotations are scoped to the member being removed:

- If a member is a project member, only that project's rotations
  are affected
- If a member is a group member, all of the projects within that
  group are affected

https://gitlab.com/gitlab-org/gitlab/-/issues/323631