Return NilMenuItems instead of nil in project menu items
In this commit we change how creating menu items in the
project sidebar works.

Before, when the user couldn't see a menu item, we will
return `nil` and avoid displaying it, but this has a problem.

If we later want to reference it, in order to add another
menu item based on that item position, the menu item
won't exist.

In this commit we return a `NilMenuItem` instead of
a `nil` value, whenever the menu item can't be rendered.