Simplify error code handling for external pipeline validation
Prior to this, the external pipeline validator would reject a pipeline
if the service returned any 4xx error code, but the behavior was
different for GitLab.com, where only a 406 error code would be used to
reject.

We simplify the handling of this error code by making 406 the only error
code to signify a rejection. This not only makes it easier to
understand, but it also makes the behavior consistent from one GitLab
instance to another.

Changelog: changed