Move endpoint to /api/v4/geo/proxy
Because `/api/v4/internal` and `/lib/api/internal/base.rb` are only
meant for GitLab Shell.

Since Workhorse doesn't know whether it's in a FOSS or EE context, we
need to implement the endpoint in FOSS to avoid 401s due to the route
not existing.

So a more appropriate location is `/api/v4/geo/proxy` and
`/lib/api/geo.rb`.

This commit also reuses the Workhorse authorization helper and Workhorse
response content type.

The affected endpoint is not used or released yet, so no changelog is
needed.