Skip download of blobs with foreign layers
Some images contains layers that might have the media type
"application/vnd.docker.image.rootfs.foreign.diff.tar.gzip"
which are not present on the registry.

In this case, the client doesn't go to the registry to fetch
the layer because it is not there but will follow the URL,
specified among other details of the layer.

Geo should not try to fetch it from the registry during sync
but skip this layer. It also does not want to follow the URL
and download it because it might violate the license.