git: Allow LFSChanges class to receive multiple revisions
The LfsChanges class currently only allows for a single revision, but in
fact Gitaly supports to get multiple revisions. We can make use of this
fact to implement the LFS pointer check in a batched manner, retrieving
all new LFS pointers which have been introduced by any of a given new
revisions.

Convert both the Gitaly client and the LfsChanges class to reflect this
such that callers can either continue to pass a single revision, or
instead pass an array of revisions.