Fix a bug displaying project commit anchors
A repository can be non-empty, yet lack a root_ref. When this is the
case, we fail to generate a link to commits, resulting in a 500 error.

Using the "default" default branch in this case allows us to generate a
link. It leads to a location that will 404, but that's a better outcome
than the current scenario.

Changelog: fixed