Add Assert that Timezone is Formatted Properly
The fixture generation of deploy freezes does not guarantee ordering, so
instead we can create a minimal map of the timezones we expect to assert
against.

This also uncovered a small issue where the timezone identifier was
somehow lost. It seems it was not used, but it would still be better to
save it instead of setting it to undefined.