Implement new active state for project sidebar context menu
In this commit we're changing the behavior of the scope or
context menu. Before, this menu only allowed us to click
on it and it'd redirect as to the `Project overview` submenu
in an active state.

That is going to change and, when we click the overview menu
it will change its state to active instead of the `Project
overview`.