Wiki: Use FindAllCommits RPC to list page versions
The wiki interface to the Git data leverages the Wiki Service in Gitaly.
This service uses LibGit2, and the Ruby sidecar. As such there's a
maintainance burden to the Wiki RPCs.

This change changes the Gitaly RPC that the `Wiki#versions` call uses to
FinddAllCommits. This RPC uses Git itself, and as such will allow the
remove of the PageVersions RPC on the Gitaly side in the next release.

Furthermore, it's expected that this change resolves other bugs too, for
example: https://gitlab.com/gitlab-org/gitlab/-/issues/330295#note_572087692
In this instance, `git-fsck(1)` can inflate a ZLib'ed Git object, where
Libgit2 for some reason fails. While I don't yet know why Libgit2 fails,
leveraging the same implementation as `git-fsck(1)` should resolve it.

Changelog: changed