Optimize read merge request permission check for references
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/271242

**Problem**

'#can_read_reference?' is an expensive call especially when we trigger
it thousands of times.

**Solution**

`:read_merge_request_iid` depends on the user and the project. This
allows us to cache the user-project pair result and apply it to all
merge requests of the same project.

Changelog: performance