Fix title, value spacing of various sidebar blocks
Spacing between title and value of sidebar blocks
is inconsistent.

This commit changes most of the sidebar blocks on
Issues and MRs to be consistent.

gitlab-utility classes are used to achieve the
same result as the `.title` class.

Changelog: changed
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/61650
EE: true