Fix sign-in when user has multiple group saml identities
If a GitLab.com user has multiple group saml identities, a
problem can occur when looking up the identity. Depending on
the order the database returns the records the wrong identity
may be chosen. This can lead to attempted creation of a duplicate
record, and ultimate error signing in.

Changelog: fixed