Flush statistics cache anytime it is updated
`Projects::UpdateStatisticsService` attempts to refresh project
statistics, but it relies the caller to flush the existing statistics
cache for the repository. If the flush doesn't happen, the refresh won't
actually do anything. For example, this might be happening in
`ProjectCacheWorker`.

We now expire the stats inside the service to ensure consistent results.

This came out of
https://gitlab.com/gitlab-org/gitlab/-/issues/255001.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/52938