Drop index with lock timeout and retry
This avoids a long running query: Instead of setting a high statement
timeout, we use rather long lock timeouts and retries in between
attempts. This should gracefully resolve the situation when a vacuum
operation is blocking the `DROP INDEX` statement.

Still, if we run out of retries, we force the execution
without lock timeout (and potentially see a long running query then).

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/329724