Clarify Group SAML user identity updates
As previously implemented the code was technically buggy and
could cause confusion in the future. That is because the call
to self.identity, if an update were necessary, would fail to
update user identities association and would not be auto-saved.
Actually, it wouldn't be possible for the identity to be found
in this manner if the extern_uid had changed. Make it clearer
by using a guard to indicate nothing is happening in this case.

Changelog: fixed