Avoid updating Elasticsearch when project is pending_delete
When a project has pending_delete set to true that means
the project is actively in the middle of being deleted. If
the Elasticsearch index is updated while that delete is
happening, a race condition occurs and the indexer will
get errors trying to load related records (such as the
project feature). This change does not initiate any
Elasticsearch updates if the project is marked as
pending_delete.

Changelog: fixed