Place multiselect drag drop behind a feature flag
The Vue library we use for the refactored boards
does not support multi-item drag&drop.

Until we complete the implementation of drag and drop
in the refactored boards, we will disable multi select
in boards by placing it behind a feature flag.

Docs change:
Add a section on 'board_multi_select'
and update Multi-select issue cards section.

Changelog: other