Added virtual scrolling to diffs
This adds virtual scrolling to the diffs app so that we can render
a lot of diff files in a very fast and performant way.
Doing virtual scrolling also helps us manage our memory usage a lot
as we aren't rendering a lot of components into the DOM.

Locally we get a lot of performance gains (close to 20s improvement
to the TBT time).

For now this is behind a feature flag. The feature flag is there so that this
can be merged but not enabled until we are sure this is something we
want.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/326773