Instrumenting project sidebar menus
In this commit, we're adding some tracking attributes
to the project sidebar. This way we can track the state of
menus and menu items when they're clicked.

In order to trigger the tracking, the menus and menu items
must have the `data-track-label` attribute. Based on that
attr and also the context of the sidebar, we calculate
the other tracking attributes.

Changelog: added