Determine Jira deployment_type based on URL
If API-based methods fail when determining the deployment_type we should
use the URL. Since Jira Cloud does not support custom domains yet we
can pretty much guess whether it's Jira Cloud or Jira Server based on
the domain. If it ends with atlassian.net then we're dealing with a Jira
Cloud, otherwise it's Jira Server.

Changelog: fixed