Set current project/group in WebHookWorker context
Currently this isn't always reliably detected, depending on how the
worker is triggered.

By setting this explicitly we always see the related project and
namespace in Kibana, and also the subscription plan in EE.

This also sets the `related_class` field to the webhook type, which
is useful for troubleshooting purposes.