Join with pending builds table to speed up queuing
This change makes it possible to perform a join with a separate table
that stores the queue of pending builds.

Instead of traversing a very large table that holds all builds that have
been processed and these that are pending too, we maintain a small table
that just stores a small amount of builds that are in the pending state.