Remove unused Namespace#all_pipelines
This is related to some work where we are looking into moving some
tables to a separate logical database. This query may be a little tricky
to do if we move `ci_pipelines` to a separate logical database.

However, it appears this method is unused anyway so we may as well just
remove it for now.