Prevent retried builds from being retried again
Previously if a build were retried, we would still allow it to be
retried again. While this is convenient for running a build again, it
also means that a build can be retried again even if there is another
build running.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/331361

Changelog: changed