Refactor tests to use group token instead of sandbox
Prior to this change, the package tests that
use the group endpoint have a sandbox runner
assigned to process the jobs at the group level.
This caused the removal of the runner to be flaky
since there can be too many runners on the general
sanbox.

This change scopes the runner to the group level
avoiding to have all the sandbox runners visible at
the project level