Ensure consistent JSON formatting across systems
This change ensures that we use a single mechanism for serializing data
to JSON (`Gitlab::Json`), rather than relying on implicit conversions
for hook data.

The root cause of this issue is that the data was sent to the background
worker as a `Hash` of data. To do so it is serialized by Sidekiq, and
then deserialized on the other end. Mostly this is fine, but in the
specific case we are fixing here (see `ci/pipeline_spec.rb`), dates were
serialized in an incompatible format to the REST API, then loaded as
strings, and never fixed.

The solution is to do the encoding _before_ we send the data to the
worker, then load the string literal and send it as is to the various
web-hook receivers. The `WebHookService` which handles both synchronous
and asynchronous execution can be made ignorant of this detail by
packaging the data in a `PrecompiledJSON` fragment, which abstracts the
serialization.

One important change is made here to `Gitlab::Json`: `#dump` is made to
respect the use of `#raw_json` methods. This is not necessary for this
particular fix, but without this change, `.dump` and `.generate` produce
different results for precompiled JSON fragments, which is not
acceptable. This is an internal `Oj` optimization, but this is only used
inside the `Gitlab::Json` module itself, and thus the abstraction does
not leak.

Changelog: fixed