Set the git default branch to "main"
In concert with the wider Git community, GitLab committed to changing
the default branch from "master" to "main" in 14.0. This change has
already happened on GitLab.com; this change enables it for
self-managed instances as well.

GitLab has support for a custom default branch name which can be used
to override this change if defaulting to "main" is inconvenient.

Unless overridden, this affects all new projects on all self-managed
GitLab instances. Existing projects will not have their default branch
changed.

Changelog: added