No Longer Send Legacy Feature Flags
Legacy feature flags have been deprecated for a while now. The beginning
of the removal process is to stop sending them entirely.

This is gated behind 2 feature flags: `remove_legacy_flags` and
`remove_legacy_flags_override`.

The former is the "main gate" to disable sending legacy feature flags.
The latter is an override control to enable for specific projects that
have not completed their migration.

Changelog: changed