Call RedisHllUserEvent only when a user is set
We're calling the `trackRedisHllUserEvent` function from parts of the
application where we can't guarantee a signed in user. The API however
assumes that there is a user set and would respond with a 401 error.

As a solution, we now only pass events to the endpoint when a user is
signed in.

This MR also fixes a potential false-positiv test by calling the
function before checking if the mocked axios client gets called.

Changelog: fixed