Use a custom regex for changelog tags
The tag format we use for gitlab-org/gitlab doesn't follow semantic
versioning; at least not as we expected. The -ee suffix we use for tags
is treated as the pre-release identifier, and when generating changelogs
we ignore pre-release tags. To solve this we need a custom tag regex
to ignore the -ee suffix.

See https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1776 for
more information.