Deprecate GraphQL PrometheusServiceID argument
Alterative commit title: "Move monitoring integrations to
`Integrations::`" (not chosen because GitLab use commit titles as their
changelog titles, and this is a `Changelog: deprecated` changelog).

This MR namespaces the "monitoring" integrations to the `Integrations::`
namespace as part of
https://gitlab.com/gitlab-org/gitlab/-/issues/201855.

As the old `PrometheusService` model was used as a GraphQL GlobalID
argument, it adds a new deprecation to allow backwards-compatibility
with clients who still use the old GlobalID string
(`"gid://gitlab/PrometheusService/<id>"`), or name
(`PrometheusServiceID`), in arguments supplied to the mutations.

https://gitlab.com/gitlab-org/gitlab/-/issues/201855
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/62645

Changelog: deprecated