Update pipeline failed notification emails to refer to jobs
The current pipeline failed notification emails refer to failures of
builds. This is confusing to some users who assume their pipeline
failed because of a problem with the 'build' job when really it could be
any job that failed.

Fix this by updating the email text to correctly refer to the failures
being those of jobs, not builds.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>

Changelog: changed