Move a mattermost file to help Rails autoload
Moving the file from `lib/mattermost/error.rb` to `lib/mattermost.rb`
aids Rails in autoloading the module and classes within it.

This was necessary after the `MattermostService` model was renamed to
`Integrations::Mattermost` and therefore the file to be named
`integrations/mattermost.rb` which Rails was attempting to autoload.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/61957#note_588568064