Add the deduplication hash to the job payload
This adds the job-deduplication hash to the payload of the job.

When the hash is part of the job, Sidekiq-server does not need to
recalculate it for the deleting it from redis. This allows us to
change the way hash is calculated without breaking deduplication on
the server-side.

It also fixes a bug that led us to investigating this, when the
serialisation of the job-arguments is different on the server-side,
the key would not correctly be deleted from Redis. This in turn would
cause the client to not enqueue new jobs for the same arguments for
the TTL of the key, even though they should be.

Changelog: fixed