Document and lint that trailers are case-sensitive
The changelog API treats Git trailers case-sensitively in various
places. We also do the same in templates when we use trailers such as
`EE` and `MR`. Supporting all this in a case-insensitive manner is not
feasible at this time.

Instead, we now document that trailers must match exactly what value is
set in the `trailer` setting. In addition, Danger now produces an error
when a changelog trailer is present but uses the wrong casing (e.g.
`changelog` instead of `Changelog`).

Changelog: other