Prevent error when accessing optional needs in pipeline graph
It is possible when building the links arrays for needs
that a defined jobs in the needs array does not exist in
the object that contains all jobs. This is because a job
with rules may or may not exist at execution time.
This commit access safely the needs or return an empty value.

Changelog: fixed