Add upcoming_reconciliations table and model
Add the upcoming_reconciliations table and UpcomingReconciliation
model.

This table is to be used for displaying alerts to customers when
their subscription will be undergoing the quarterly reconciliation
process.

The table will hold the date when the next reconciliation is to
occur, and a date when an alert should start being displayed to the
customer. This information will be received from CustomersDot.

The table will hold a row for each relevant namespace in SaaS,
which should be about 9,000 - 10,000 rows (at time of commit). This
is because customers buy subscriptions for top-level namespaces in
SaaS.

In self-managed, the table should hold one row only, because the
entire self-managed instance uses one subscription/license.

MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/63054

Changelog: added