Cache commit mentions in the database
When a commit is mentioned in a note, we already store that fact in the
database. However, when a commit mentions an issue in a note, we don't
store that in the database right now. Start doing so.

The table we store this information in is keyed on the commit ID, which
is the same across all projects, so we don't have problems with the same
commit appearing in multiple projects in the same way we do for cross-
references.

Changelog: added