Move Labkit::Middleware::Rack to the top
Labkit context is essential for our application so that it abtracts
context data passing in different layers of a request lifecyle. Before
this change, there is a middleware that initiates context and setup
collaboration ID. That middleware is unfornuately in the middleware of
the middleware stack. Moving it to the top, right after RequestId
middleware, unlocks some useful use cases for us.

Issue: https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/1083
Change: changed