Update the vendored cluster management Project Template
We've decided to give the users the full power and responsibility of
managing their own Helmfiles. For this change, will use the
cluster-applications:v1.0.0+ which won't have the chart built-in anymore.

Those configurations related to the config.yaml file and all that logic
to allow the merge of values got removed from those Helmfiles. This
logic is not needed anymore since the files are no longer locked inside
the docker image and the user has full control of them over their
project template repository.

Also, it's worth noting that we're removing Ingress WAF which is
deprecated.

Another important point is that the Helmfiles being included here were
tested to work only against Helm3, which is expected from the
cluster-applications v1.0.0.

Some other work implemented here:

- Set all applications install default to true
   but the main Helmfile comments them out to not
   accidentally install/update anything.

- Use helm stable instead of gitlab stable archive.

- Use default branch instead of master for the CI.

- Default namespace to gitlab-managed-apps as a
   legacy to facilitate migration from GMA v1 and v2.

- Don't add a default insecure password to sentry DB.

- Remove missing file handlers
   Those are not needed anymore since we don't have logic
   over the existence of those helmfiles and the values
   files were all created by default even if blank.

- Update default runner image to ubuntu:20.04.

- Adds script to adopt CM CRDs and ClusterIssuer when the
user had previously installed CM through the UI.

- Vault chart is upgraded to 0.5.0 which is a necessary
   breaking change to comply with Helm v3.

Changelog: changed