Document zero downtime Praefect elector migration steps
We currently recommend shutting down all of the Praefect nodes when
switching to repository-specific primaries in Praefect. This leads to
short downtime until the Praefect. It is possible to avoid the downtime
by taking a few more steps during the migration. This commit documents
the steps to do a zero downtime migration to repository-specific
primaries.