Check existence of export object and report status to user
Now that the project export status in the database may be updated before
the upload has transferred to object storage, we should inform the user
that the export is still not available. We do this when the user
attempts to request the actual download since the existence check
requires a HEAD request to ensure the file is present.