Prevent CI schema from registering if schema_linting flag is off
If the schema_linting flag is off, then an error is returned
when trying to register the schema. This is intended to prevent
self hosted instances from accessing 3rd parties ressources
such as the CI schema.

This commit now wrap the registration of the CI schema
inside a conditional to only executes when schema_linting
is turned on.

Changelog: fixed