Show update branch message when MR branch diverged from target branch
Previously, the staleness message shown for the security report on an MR
was determined using this (pseudo-code) logic:

    if target_branch_security_report_is_older_than_7_days:
      if mr_branch_diverged_from_target_branch:
        render "Update your branch with latest changes from target branch"
      else:
        render "Run a new pipeline for target branch"

This change un-nests the logic so that each condition (whether the
target branch's latest security report is older than 7 days; whether the
MR branch has diverged from the target branch) alone produces the
relevant warning message. That is, the logic has changed to this:

    if mr_branch_diverged_from_target_branch:
      render "Update your branch with latest changes from target branch"
    else if target_branch_security_report_is_older_than_7_days:
      render "Run a new pipeline for target branch"

This means that the mr_branch_diverged_from_target_branch warning will
now almost always be shown for highly active projects. This is because
the security report for an MR is based on the comparison of the security
report of the latest HEAD of the target branch and the HEAD of the MR
branch, and therefore won't give entirely correct results if the two
have diverged.

The target_branch_security_report_is_older_than_7_days message will
continue to be shown with the same frequency.

This change also aligns the implementation with the documented[doc]
behaviour.

While this is now more correct, it may be a bit more noisy than the
previous behaviour. There is future work planned[1] to use the merge
base for the comparison instead, which would improve this further.

Finally, this also:
 - refactors the component tests to conform a bit more to our current
   standards;
 - replaces some Bootstrap utility classes with GitLab UI classess.

Addresses part of https://gitlab.com/gitlab-org/gitlab/-/issues/267504.

[1]: https://gitlab.com/gitlab-org/gitlab/-/issues/295167
[doc]: https://docs.gitlab.com/13.12/ee/user/application_security/index.html#outdated-security-reports

Changelog: changed
EE: true