Fix CI editor template not updating on commit
When the default content is loaded for the pipeline
editor, there was an issue where changing the content
of the file would not be saved when you commit. This
was due to using a computed prop to store the content:
if it was a new file, we were using the template,
and if not, we were using the normal content.
However when the user starts typing, it is
still a new Ci config file, so the content
was never properlyupdated in hte global store.

Changelog: fixed