Create close issue event on issue move
When an issue is moved to a different project we create a
copy of current issue and close current issue. We need to also
create a close issue resource event so that that issue being
closed is also reflected in burndown/burnup charts that rely
on resource state events.

Changelog: fixed
EE: true