Created pre-renderer and scroller sync
The pre-renderer pre-renders the diff files one by one to catch
the sizes of the diff file.
This is important for when we want to scroll to a set index
as without this the virtual scroller will try to calculate the sizes of each
one before the specified index which can lock the page.

The scroll sync component is used to sync up virtual scroller
scroll to index with whatever diff file the user wants to be viewing.

There will still be some lock-ups and maybe some slowness at certain
points, but in order to get over them hurdles we need to improve
the individual diff file components.