Enable deep linking on environments page for tabs
When migrating to `GlTabs`, no syncing with the query parameters was
done.

We correct that by adding an `activeTab` computed property that searches
for the tab selected by the query parameters, defaulting to the first
tab if none is found. This is bound to the `GlTabs`' `:value` property.

As the query parameter is always updated on tab selection, changing tabs
works as expected.

Changelog: fixed