Remove old cloud licenses upon sync or activation
Fixes a bug in the SyncSeatLinkRequestWorker where the current license
was not being deleted despite being tested.  This appears to be a
problem with the way License.current uses caching.  The .destroy!
command does not actually delete the record from the DB.  The fix
instead deletes all existing cloud licenses uses delete_all.

Applies a similar fix for removing old cloud licenses to the
ActivateService.  Also fixes a minor bug where new licenses created in
the ActivateService did not create a record with the cloud attribute.