Proper fix for order-dependent migration spec failure
It turned out that we have hooks for :migration specs to re-up the
migrations after the specs have run. It's just that in this case, the
schema: was set in-correctly so this migration was not downed in the
before hook.

Because this migration was not downed, that means the migration will not
be upped in the after hook. So we fix the schema: to a version number
slightly before this migration