gitaly_client: Implement support for `#list_blobs`
The new `#list_blobs` RPC allows the caller to enumerate blobs via a set
of revisions. It's thus a more flexible sibling to `#get_new_blobs`,
which only allows a single positive revision and will automatically add
`--not --all`.

Add support for this new RPC via a new `#list_blobs` function in the
`BlobService`.