Show warning when files are hidden in an MR
What was happening was that, for example, if our limit files was 1,000
and our diff had 1,500 we would only save 1000 files from gitaly in
the database, and when reading them back it wouldn't overflow since it
only contains 1,000 files, and as such, no warning would be shown.
However, it is still overflown, and the warning should be shown.

MR: gitlab.com/gitlab-org/gitlab/-/merge_requests/64515
Issue: gitlab.com/gitlab-org/gitlab/-/issues/332883

Changelog: fixed