Fix infrastructure menu link to the first visible menu item
Current instratructure menu link is harcoded to the kubernetes
link. Nevertheless, if the kubernetes menu item can't be displayed
because the user doesn't have enough access rights, the main
menu still points to it.

We need to change that and always point to the first visible
menu item.

Changelog: fixed