Draft: Add jobs failed alert to security MR widget
This adds the alert, as originally designed in
https://gitlab.com/gitlab-org/gitlab/-/issues/12896, to the security MR
widget.

The design is being rethought at the moment, perhaps in large part due
to https://gitlab.com/groups/gitlab-org/-/epics/5710, which is an effort
to redesign all merge request widgets.

This alert implementation is almost certainly not going to be part of
the eventual design (as far as I can tell), but the general structure of
how to implement it is here (minus tests).

Note that several unrelated GraphQL queries were modified (id fields
were added to various entities) due to a manifestation of
https://gitlab.com/gitlab-org/gitlab/-/issues/326101.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/267504.