Draft: Show security MR widget in edge case
Specifically, this makes the security MR widget show when the source
branch is based on an old commit of the target branch, such that it
isn't configured to run any security scans. Meanwhile, the target branch
has advanced and _is_ configured to run security scans. This gives an
opportunity to prompt the user to rebase the source branch, in order to
run security scans.

It should be noted that this is a rare edge case, and I'm not sure how
worthwhile it is to actually handle it. A project is only likely to
enable security scanning _once_, and once it has, any new branches
created after it are unaffected. Only a few existing MRs, based
on an old commit of the target branch, would show this, until they're
rebased. To put this another way, the window when this is likely to
happen for a given project is brief, and happens probably only once in
a given project's lifetime.

This minimal implementation exposes the `security_reports_up_to_date?`
field to the frontend, as per the second example in
https://gitlab.com/gitlab-org/gitlab/-/issues/275818.

Unfortunately, more work need to be done, as the MR widget renders in
this case with a successful status, and message "Security scanning
detected no vulnerabilities" which is not correct - security scanning
did not even run!

Exactly _what_ should be rendered instead varies from design to design,
so it's not clear what to actually implement right now. As with the
previous commit, this may become clearer with
https://gitlab.com/groups/gitlab-org/-/epics/5710.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/275818