Stop logging response field in Lograge
We had a `response` field in our logs from the ApplicationController,
which had some custom logging for the body of 422 responses to JSON
requests. We removed the custom logging in the controller, but not from
the Lograge configuration. This didn't cause a problem because the field
wasn't present and so wasn't logged, but Rails 6.1 added its own
`response` field to the log:
https://github.com/rails/rails/commit/b5c6f33f0d301d8766acb5e516defe4e4eae3119

That doesn't have a convenient string representation by default, which
meant we were logging things like this:

    "response": "#<ActionDispatch::Response:0x00007fcb384bc8b8>"