Change modal error message for restricted emails
Displays a custom error message in the Invite members modal for
the API error message returned when an email address is invited
that is not on a restricted domain list. The message is different
depending on whether it is an invite to a group or to a project,
and whether the page viewer is able to administer the access
restrictions or not.