Add validate account button on CI/CD shared runners page
Previously once a user disabled shared runners, the only way to get this
modal to display would be to create a new project with a
`.gitlab-ci.yml`, let the pipeline fail, visit the pipeline, and then
find this same modal.

If credit card validation is required to enable shared runners for a
given project, this change will now display the `Validate account`
button and direct the user to the validation.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/334322

Changelog: added
EE: true