Log and advise about push-rule-rejected commit
Because the GitLab Shell log doesn't include another detail
that references other GitLab components [1],
it is very difficult to find out where and
why exactly a certain push rule was broken.

This commit adds the commit's SHA into that output
and suggests a next step to the user.

[1] https://gitlab.com/gitlab-org/gitlab/-/blob/14-0-stable-ee/doc/administration/logs.md#gitlab-shelllog

Changelog: changed
EE: true